_COMMUNICATION/team_40/legacy-media-index-50-2026-04-11/requirements-semantic.txt
# שכבה סמנטית (CLIP) — התקנה חד-פעמית; גודל משוער: PyTorch + משקלים
torch>=2.1
open_clip_torch>=2.24.0
torchvision>=0.16
